Fetch Zendesk articles as Markdown
