+++ backend/app/core/utils/web_scraping_utils.py
@@ -1,6 +1,5 @@
 import asyncio
 import ipaddress
-import re
 import socket
 from urllib.parse import urljoin, urlparse
 
@@ -22,7 +21,6 @@ _FORWARDED_HEADER_ALLOWLIST = frozenset({
     "if-none-match",
 })
 
-_LINK_TITLE_RE = re.compile(r'(\]\(\S+?)\s+"[^"]*"(\))')
 
 def _is_blocked_ip(addr: str) -> bool:
     try:
@@ -123,25 +121,3 @@ async def fetch_from_url(
         html = await page.content()
         await browser.close()
         return html
-
-
-def _normalize_whitespace(s: str) -> str:
-    s = re.sub(r"[ \t]+\n", "\n", s)   # strip trailing spaces before newlines
-    s = re.sub(r"\n{3,}", "\n\n", s)   # collapse 3+ blank lines to 2
-    s = re.sub(r"[ \t]{2,}", " ", s)   # collapse runs of spaces/tabs (does NOT touch newlines)
-    return s.strip()
-
-
-def _strip_link_titles(md: str) -> str:
-    return _LINK_TITLE_RE.sub(r"\1\2", md)
-
-
-def html2markdown(html: str, base_url: str | None = None) -> str:
-    import html2text as _html2text
-
-    # baseurl resolves relative hrefs to absolute URLs
-    h = _html2text.HTML2Text(baseurl=base_url or "")
-    h.ignore_links = False   # keep links as [text](url)
-    h.ignore_images = True
-    h.body_width = 0         # no hard wraps
-    return _normalize_whitespace(_strip_link_titles(h.handle(html or "")))
